Test 'inteldfp' symbol visibility only on Linux.
The IBM platform fails to build the intel test driver with new
symbol visibility tests.  This is due to a spurious redefnition of
'NAN'.  These tests are now performed only on the Linux platform.
This should not be a problem , as we're testing the viability of a
compiler, and not the functionality of a component.